+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/__init__.py
@@ -1,7 +1,7 @@
 from .electricity_market.child_metering_points.wrapper import ChildMeteringPoints
 from .electricity_market.consumption_metering_point_periods.wrapper import ConsumptionMeteringPointPeriods
 from .electricity_market.repository import Repository as ElectricityMarketRepository
-from .measurements.measurements_bronze.wrapper import MeasurementsBronze
+from .measurements.measurements_calculated.wrapper import MeasurementsBronze
 from .measurements.measurements_gold.wrapper import TimeSeriesPoints
 from .measurements.repository import Repository as MeasurementsRepository
 from .spark_initializor import initialize_spark
