--- source/electrical_heating/src/opengeh_electrical_heating/infrastructure/electricity_market/__init__.py
@@ -1,5 +1,0 @@
-from .repository import Repository
-
-__all__ = [
-    "Repository",
-]
