--- source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements/measurements_bronze/database_definitions.py
@@ -1,3 +1,0 @@
-class MeasurementsBronzeDatabase:
-    DATABASE_NAME = "measurements_bronze"
-    MEASUREMENTS_NAME = "measurements"
