+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements/measurements_calculated/database_definitions.py
@@ -1,0 +1,3 @@
+class MeasurementsCalculatedDatabase:
+    DATABASE_NAME = "measurements_calculated"
+    MEASUREMENTS_NAME = "measurements"
