+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements/measurements_calculated/wrapper.py
@@ -1,18 +1,18 @@
 from pyspark.sql import DataFrame
 from pyspark_functions.data_frame_wrapper import DataFrameWrapper
 
-from opengeh_electrical_heating.infrastructure.electricity_market.consumption_metering_point_periods.schema import (
-    consumption_metering_point_periods_v1,
+from opengeh_electrical_heating.infrastructure.measurements.measurements_calculated.schema import (
+    measurements_bronze_v1,
 )
 
 
-class ConsumptionMeteringPointPeriods(DataFrameWrapper):
-    """Represents the consumption metering point periods data structure."""
+class MeasurementsBronze(DataFrameWrapper):
+    """Represents the bronze measurements data structure."""
 
     def __init__(self, df: DataFrame):
         super().__init__(
             df,
-            consumption_metering_point_periods_v1,
+            measurements_bronze_v1,
             # We ignore_nullability because it has turned out to be too hard and even possibly
             # introducing more errors than solving in order to stay in exact sync with the
             # logically correct schema.
