+++ source/electrical_heating/tests/job_tests/test_electrical_heating_job.py
@@ -8,8 +8,8 @@ from pyspark.sql import functions as F
 
 from opengeh_electrical_heating.application import execute_application
 from opengeh_electrical_heating.application.job_args.environment_variables import EnvironmentVariable
-from opengeh_electrical_heating.infrastructure.measurements.measurements_bronze.database_definitions import (
-    MeasurementsBronzeDatabase,
+from opengeh_electrical_heating.infrastructure.measurements.measurements_calculated.database_definitions import (
+    MeasurementsCalculatedDatabase,
 )
 
 
@@ -35,6 +35,6 @@ def test_execute_with_deps(spark: SparkSession, job_environment_variables: dict,
 
     # Assert
     actual = spark.read.table(
-        f"{MeasurementsBronzeDatabase.DATABASE_NAME}.{MeasurementsBronzeDatabase.MEASUREMENTS_NAME}"
+        f"{MeasurementsCalculatedDatabase.DATABASE_NAME}.{MeasurementsCalculatedDatabase.MEASUREMENTS_NAME}"
     ).where(F.col("orchestration_instance_id") == orchestration_instance_id)
     assert actual.count() > 0
